backend/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["savosinvitali <savosinvitali@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
pydantic-settings = "^2.1.0"
psycopg = {extras = ["binary"], version = "^3.1.17"}
alembic = "^1.13.1"
pip = "^23.3.2"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
install = "^1.3.5"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.7"



[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.5"
pytest-dotenv = "^0.5.2"


[tool.pytest.ini_options]
pythonpath = [
  ".", "src",
]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
